refactor(atelier): split pragma codegen helpers
