+++ crates/vize_atelier_jsx/src/compile.rs
@@ -3,12 +3,12 @@
 //! The module is lowered once, then each render root is routed to VDOM, Vapor,
 //! or SSR according to the configured default and any directive prologue.
 
+mod babel;
 mod component;
 mod render_exports;
 #[cfg(test)]
 mod tests;
 
-use oxc_allocator::Allocator;
 use vize_carton::{Bump, FxHashSet, String};
 use vize_croquis::Croquis;
 
@@ -19,6 +19,8 @@ use crate::ssr::compile_lowered_root_to_ssr;
 use crate::vapor::{VaporCompileOptions, compile_root_to_vapor};
 use crate::vdom::{VdomCompatOptions, VdomCompileOptions, compile_root_to_vdom};
 use crate::{JsxLang, JsxOutputMode, lower_source_with_compat};
+
+use self::babel::{collision_free_transform_on_helper, resolve_vnode_factory};
 
 pub use component::JsxComponent;
 
@@ -274,22 +276,11 @@ pub fn compile_jsx_with_babel_pragma(
             .roots
             .iter()
             .any(|root| resolve_mode(root.mode, config.default_mode) == JsxOutputMode::Vdom);
-    let vnode_factory = pragma
-        .map(str::trim)
-        .filter(|pragma| !pragma.is_empty())
-        .filter(|_| config.compat.is_babel() && has_vdom_root)
-        .and_then(|pragma| {
-            if valid_pragma_expression(pragma) {
-                Some(pragma)
-            } else {
-                diagnostics.push(JsxDiagnostic::error(
-                    "Babel JSX pragma must be a valid JavaScript expression",
-                    0,
-                    0,
-                ));
-                None
-            }
-        });
+    let vnode_factory = resolve_vnode_factory(
+        pragma,
+        config.compat.is_babel() && has_vdom_root,
+        &mut diagnostics,
+    );
 
     // Move the analysis into the arena so the transforms can borrow it.
     let analysis: &Croquis = &*bump.alloc(lowered.analysis);
@@ -356,24 +347,4 @@ pub fn compile_jsx_with_babel_pragma(
         source: String::from(source),
         diagnostics,
     }
-}
-
-fn valid_pragma_expression(pragma: &str) -> bool {
-    let mut probe = String::from("const __vize_pragma = (");
-    probe.push_str(pragma);
-    probe.push_str(");");
-    let allocator = Allocator::default();
-    !crate::parse_module(&allocator, probe.as_str(), JsxLang::Jsx).has_errors()
-}
-
-/// Babel allocates a fresh helper binding when the source already declares
-/// `_transformOn`. A conservative source-text check gives the generated module
-/// the same collision safety without coupling this option to a second semantic
-/// analysis pass.
-fn collision_free_transform_on_helper(source: &str) -> String {
-    let mut helper = String::from("_transformOn");
-    while source.contains(helper.as_str()) {
-        helper.push('_');
-    }
-    helper
 }
